pyproject.toml
[tool.poetry]
name = "search-handwritten"
version = "0.1.0"
description = "An API to search handwritten applications. Uses VLM as OCR model and EmbedAnything for embedding generation and Qdrant for vector database"
authors = ["Akshay Ballal <arballal95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
qdrant-client = "^1.12.1"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
embed-anything = "^0.5.2"
qwen-vl-utils = "^0.0.8"
streamlit = "^1.40.2"
pillow = "^11.0.0"
lancedb = "^0.18.0"
transformers = "^4.48.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
accelerate = "^1.3.0"
streamlit-file-browser = "^3.2.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

